+++ Editor/SkinningModule/IMGUI/ISkeletonView.cs
@@ -29,20 +29,22 @@ namespace UnityEditor.U2D.Animation
 
     internal interface ISkeletonView
     {
-        int InvalidID { get; set; }
+        EntityId InvalidID { get; set; }
         SkeletonMode mode { get; set; }
         int defaultControlID { get; set; }
-        int hoveredBoneID { get; }
-        int hoveredJointID { get; }
-        int hoveredBodyID { get; }
-        int hoveredTailID { get; }
-        int hotBoneID { get; }
+
+        EntityId hoveredBoneID { get; }
+        EntityId hoveredJointID { get; }
+        EntityId hoveredBodyID { get; }
+        EntityId hoveredTailID { get; }
+        EntityId hotBoneID { get; }
+
         void BeginLayout();
         void EndLayout();
         bool CanLayout();
         Vector3 GetMouseWorldPosition(Vector3 planeNormal, Vector3 planePosition);
-        void LayoutBone(int id, Vector3 position, Vector3 endPosition, Vector3 forward, Vector3 up, Vector3 right, bool isChainEnd);
-        bool DoSelectBone(out int id, out bool additive);
+        void LayoutBone(EntityId id, Vector3 position, Vector3 endPosition, Vector3 forward, Vector3 up, Vector3 right, bool isChainEnd);
+        bool DoSelectBone(out EntityId id, out bool additive);
         bool DoRotateBone(Vector3 pivot, Vector3 normal, out float deltaAngle);
         bool DoMoveBone(out Vector3 deltaPosition);
         bool DoFreeMoveBone(out Vector3 deltaPosition);
@@ -51,7 +53,7 @@ namespace UnityEditor.U2D.Animation
         bool DoChangeLength(out Vector3 endPosition);
         bool DoCreateBoneStart(out Vector3 position);
         bool DoCreateBone(out Vector3 position);
-        bool DoSplitBone(out int id, out Vector3 position);
+        bool DoSplitBone(out EntityId id, out Vector3 position);
         bool DoRemoveBone();
         bool DoCancelMultistepAction(bool force);
         bool IsActionActive(SkeletonAction action);