+++ Runtime/BatchedDeformation/DeformationJobs.cs
@@ -32,8 +32,8 @@ namespace UnityEngine.U2D.Animation
         public int tangentVertexOffset;
         public int deformVerticesStartPos;
         public int previousDeformVerticesStartPos;
-        public int transformId;
-        public NativeCustomSlice<int> boneTransformId;
+        public EntityId transformId;
+        public NativeCustomSlice<EntityId> boneTransformId;
         public NativeCustomSlice<Bounds> boneBounds;
     }
 
@@ -72,9 +72,9 @@ namespace UnityEngine.U2D.Animation
         [ReadOnly]
         public NativeArray<SpriteSkinData> spriteSkinData;
         [ReadOnly]
-        public NativeHashMap<int, TransformAccessJob.TransformData> rootTransformIndex;
+        public NativeHashMap<EntityId, TransformAccessJob.TransformData> rootTransformIndex;
         [ReadOnly]
-        public NativeHashMap<int, TransformAccessJob.TransformData> boneTransformIndex;
+        public NativeHashMap<EntityId, TransformAccessJob.TransformData> boneTransformIndex;
         [WriteOnly]
         public NativeArray<float4x4> finalBoneTransforms;
 
@@ -83,7 +83,7 @@ namespace UnityEngine.U2D.Animation
             int x = boneLookupData[i].x;
             int y = boneLookupData[i].y;
             SpriteSkinData ssd = spriteSkinData[x];
-            int v = ssd.boneTransformId[y];
+            EntityId v = ssd.boneTransformId[y];
             int index = boneTransformIndex[v].transformIndex;
             if (index < 0)
                 return;