+++ Runtime/NativeArrayHelper.cs
@@ -33,7 +33,7 @@ namespace UnityEngine.U2D.Animation
             nativeArray = newArray;
         }
 
-        public static void DisposeIfCreated<T>(this NativeArray<T> nativeArray) where T : struct
+        public static void DisposeIfCreated<T>(this ref NativeArray<T> nativeArray) where T : struct
         {
             if (nativeArray != default && nativeArray.IsCreated)
                 nativeArray.Dispose();